+++ baibaibot/kapi.py
@@ -118,12 +118,12 @@ class KAPI:
         :type data: dict
         :param headers: (optional) HTTPS headers
         :type headers: dict
-        :param timeout: (optional) if not ``None``, a :py:exc:`requests.HTTPError`
+        :param timeout: (optional) if not ``None``, a `requests.HTTPError`
                         will be thrown after ``timeout`` seconds if a response
                         has not been received
         :type timeout: int or float
         :returns: :py:meth:`requests.Response.json`-deserialised Python object
-        :raises: :py:exc:`requests.HTTPError`: if response status not successful
+        :raises: `requests.HTTPError`: if response status not successful
 
         """
         if data is None:
@@ -152,7 +152,7 @@ class KAPI:
         :type method: str
         :param data: (optional) API request parameters
         :type data: dict
-        :param timeout: (optional) if not ``None``, a :py:exc:`requests.HTTPError`
+        :param timeout: (optional) if not ``None``, a `requests.HTTPError`
                         will be thrown after ``timeout`` seconds if a response
                         has not been received
         :type timeout: int or float
