+++ sample-config.json
@@ -1,30 +1,59 @@
 {
-    "fee_percent": 0.12,
     "loglevel": "INFO",
-    "order_lifespan_seconds": 900,
-    "inter_apireq_sleep_seconds": 0.1,
-    "markets": [
-        {
-            "pair": "XXLMZEUR",
-            "sell": {
-                "order_count": 3,
-                "pcnt_bump_a": 0.25,
-                "pcnt_bump_c": 0.5,
-                "rebuy_bump_percent": 2.5,
-                "vol_percent": 99.0
-            },
-            "buy": {
-                "order_count": 3,
-                "pcnt_bump_a": 0.25,
-                "pcnt_bump_c": 0.5,
-                "resell_bump_percent": 3.0,
-                "vol_percent": 50.0
-            },
-            "min": {
-                "ask": 0.01,
-                "bid": 0.01,
-                "vwap": 0.01
-            }
+    "exchanges": {
+        "gate.io": {
+            "markets": [
+                {
+                    "pair": "VEGA_USDT",
+                    "sell": {
+                        "order_count": 3,
+                        "pcnt_bump_a": 0.25,
+                        "pcnt_bump_c": 0.5,
+                        "rebuy_bump_percent": 2.5,
+                        "vol_percent": 99.0
+                    },
+                    "buy": {
+                        "order_count": 3,
+                        "pcnt_bump_a": 0.25,
+                        "pcnt_bump_c": 0.5,
+                        "resell_bump_percent": 3.0,
+                        "vol_percent": 50.0
+                    },
+                    "min": {
+                        "ask": 0.01,
+                        "bid": 0.01,
+                        "vwap": 0.01
+                    }
+                }
+            ]
+        },
+        "Kraken": {
+            "fee_percent": 0.12,
+            "order_lifespan_seconds": 900,
+            "inter_apireq_sleep_seconds": 0.1,
+            "markets": [
+                {
+                    "pair": "XXBTZEUR",
+                    "sell": {
+                        "order_count": 2,
+                        "pcnt_bump_a": 0.75,
+                        "pcnt_bump_c": 10.0,
+                        "rebuy_bump_percent": 2.5,
+                        "vol_percent": 10.0
+                    },
+                    "buy": {
+                        "order_count": 0,
+                        "pcnt_bump_a": 0.75,
+                        "pcnt_bump_c": 1.0,
+                        "resell_bump_percent": 2.5,
+                        "vol_percent": 0.0
+                    },
+                    "min": {
+                        "ask": 10000,
+                        "bid": 10000
+                    }
+                }
+            ]
         }
-    ]
+    }
 }
