+++ sample-keys.json
@@ -1,0 +1,11 @@
+{
+    "gate.io": {
+        "host": "https://api.gateio.ws/api/v4",
+        "key": "api_key",
+        "secret": "api_secret"
+    },
+    "Kraken": {
+        "key": "api_key",
+        "secret": "api_secret"
+    }
+}
